Refactored frontend components into separate files
